twitter kadai

+++ app/Http/Controllers/TasksController.php
@@ -9,13 +9,22 @@ use App\Task;
 class TasksController extends Controller
 {
     
-     //getでtasks/にアクセスされた場合の一覧表示処理
     public function index()
     {
-        $tasks = Task::all();
-        
-        return view("tasks.index",["tasks" => $tasks,
-        ]);
+        $data = [];
+        if (\Auth::check()) { // 認証済みの場合
+            // 認証済みユーザを取得
+            $user = \Auth::user();
+            // ユーザの投稿の一覧を作成日時の降順で取得
+            $tasks = $user->tasks()->orderBy('created_at', 'desc')->paginate(10);
+
+            $data = [
+                'user' => $user,
+                'tasks' => $tasks,
+            ];
+        }
+        // Welcomeビューでそれらを表示
+        return view('welcome', $data);
     }
 
     //getでtasks/createにアクセスされた場合の新規登録画面表示処理
@@ -36,11 +45,12 @@ class TasksController extends Controller
             "status" => "required|max:10",
             "content" => "required|max:255",
             ]);
-        //タスクを作成
-        $task = new Task;
-        $task->status = $request->status;
-        $task->content = $request->content;
-        $task->save();
+        
+        //認証済みユーザの投稿として作成（リクエストされた値を元に作成）
+        $request->user()->tasks()->create([
+            "content" => $request->content,
+            "status" => $request->status,
+            ]);
         
         //トップページへリダイレクト
         return redirect("/");
@@ -51,11 +61,17 @@ class TasksController extends Controller
     {
         //idの値でタスクを検索して取得
         $task = Task::findOrFail($id);
+        if (\Auth::id() === $task->user_id){
         
         //メッセージ詳細ビューでそれを表示
         return view("tasks.show",[
             "task" => $task,
             ]);
+        
+        }
+        else{
+            return redirect("/");
+        }
     }
 
     //getでtasks/(id)/editにアクセスされた場合の更新画面表示処理
@@ -81,10 +97,13 @@ class TasksController extends Controller
             ]);
         //idの値でタスクを検索して取得
         $task = Task::findOrFail($id);
-        //タスクを更新
-        $task->status = $request->status;
-        $task->content = $request->content;
-        $task->save();
+        //認証済みユーザがその投稿の所有者である場合は、投稿を更新
+        if (\Auth::id() === $task->user_id){
+            $task->status = $request->status;
+            $task->content = $request->content;
+            $task->save(); 
+        }
+        
         
         //トップページへリダイレクト
         return redirect("/");
@@ -95,8 +114,10 @@ class TasksController extends Controller
     {
         //idの値でタスクを検索して取得
         $task = Task::findOrFail($id);
-        //タスクを削除
-        $task->delete();
+        //認証済みユーザがその投稿の所有者である場合は、投稿を削除
+        if (\Auth::id() === $task->user_id){
+            $task->delete();
+        }
         
         //トップページへリダイレクト
         return redirect("/");
